requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
dialogflow==1.1.0
Django==3.1.2
djangorestframework==3.12.1
google-api-core==1.23.0
google-auth==1.23.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
idna==2.10
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyTelegramBotAPI==3.7.3
pytz==2020.1
requests==2.25.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.1
